+++ effectful/ops/syntax.py
@@ -1,18 +1,8 @@
 import collections
 import dataclasses
 import functools
-import inspect
 import typing
-from typing import (
-    Annotated,
-    Callable,
-    Generic,
-    Mapping,
-    Optional,
-    Sequence,
-    Type,
-    TypeVar,
-)
+from typing import Annotated, Callable, Generic, Optional, Type, TypeVar
 
 import tree
 from typing_extensions import Concatenate, ParamSpec
@@ -393,83 +383,24 @@ def _(value: T) -> T:
 
 
 @defdata.register(object)
-class _BaseTerm(Generic[T], Term[T]):
-    _op: Operation[..., T]
-    _args: Sequence[Expr]
-    _kwargs: Mapping[str, Expr]
+def _(op, *args, **kwargs):
+    from effectful.internals.base_impl import _BaseTerm
 
-    def __init__(
-        self,
-        op: Operation[..., T],
-        *args: Expr,
-        **kwargs: Expr,
-    ):
-        self._op = op
-        self._args = args
-        self._kwargs = kwargs
-
-    def __eq__(self, other) -> bool:
-        return syntactic_eq(self, other)
-
-    @property
-    def op(self):
-        return self._op
-
-    @property
-    def args(self):
-        return self._args
-
-    @property
-    def kwargs(self):
-        return self._kwargs
+    return _BaseTerm(op, *args, **kwargs)
 
 
 @defdata.register(collections.abc.Callable)
-class _CallableTerm(Generic[P, T], _BaseTerm[collections.abc.Callable[P, T]]):
-    def __call__(self, *args: Expr, **kwargs: Expr) -> Expr[T]:
-        from effectful.ops.semantics import call
+def _(op, *args, **kwargs):
+    from effectful.internals.base_impl import _CallableTerm
 
-        return call(self, *args, **kwargs)  # type: ignore
+    return _CallableTerm(op, *args, **kwargs)
 
 
 @defterm.register(collections.abc.Callable)
-def _(value: Callable[P, T]):
-    from effectful.internals.runtime import interpreter
-    from effectful.ops.semantics import apply, call
+def _(value: Callable[P, T]) -> Expr[collections.abc.Callable[P, T]]:
+    from effectful.internals.base_impl import _unembed_callable
 
-    assert not isinstance(value, Term)
-
-    try:
-        sig = inspect.signature(value)
-    except ValueError:
-        return value
-
-    for name, param in sig.parameters.items():
-        if param.kind in (
-            inspect.Parameter.VAR_POSITIONAL,
-            inspect.Parameter.VAR_KEYWORD,
-        ):
-            raise NotImplementedError(
-                f"cannot unembed {value}: parameter {name} is variadic"
-            )
-
-    bound_sig = sig.bind(
-        **{name: defop(param.annotation) for name, param in sig.parameters.items()}
-    )
-    bound_sig.apply_defaults()
-
-    with interpreter(
-        {
-            apply: lambda _, op, *a, **k: defdata(op, *a, **k),
-            call: call.__default_rule__,
-        }
-    ):
-        body = value(
-            *[a() for a in bound_sig.args],
-            **{k: v() for k, v in bound_sig.kwargs.items()},
-        )
-
-    return deffn(body, *bound_sig.args, **bound_sig.kwargs)
+    return _unembed_callable(value)
 
 
 def syntactic_eq(x: Expr[T], other: Expr[T]) -> bool:
